chore: 아바타 엔드포인트 통합
